refactor(mocks): scaffold mockApi/ directory + shared.ts leaf (#991 M2)
Pure rename: mockApi.ts -> mockApi/index.ts with its own relative
specifiers repointed one level deeper, plus a shared.ts leaf holding
jitter/delay/requireSession/mockScenario/oidcDemo and the users roster
(D5). All nine importers and the dynamic import("./mockApi") resolve to
the directory index unchanged. No behaviour change.
